+++ CHANGELOG.md
@@ -14,7 +14,9 @@ and this project adheres to [Semantic Versioning](http://semver.org).
 - Changed - cli.exec() returns was work attempted, instead of exit code
 - Removed - cli.parse() and cli.exec() variants taking an ostream. Use 
     with cli.printError() instead.
-- Changed - cli.fail() now returns void instead of false.
+- Changed - cli.fail() and cli.badUsage() now returns void instead of false
+- Added - cli.parseExit() for aborting parsing w/o error
+- Changed - all types of application actions now return void
 
 ## dimcli 6.2.1 (2022-02-07)
 - Fixed - minWidth of 0 rejected in text columns